Fixed suggested fixes from code review
