+++ src/js/Book.js
@@ -4,35 +4,12 @@
  ******************************************************************************/
  const React = require('react');
 
- var Verse = React.createClass({
-   render: function() {
-     return (
-    <p key={this.props.id}>
-      <strong>{this.props.verseNumber} </strong>
-      {this.props.verseText}
-    </p>
-  );
-   }
- });
- var Chapter = React.createClass({
-   render: function() {
-     return (
-    <div>
-        <h5 key={this.props.chapterNum}>
-        <strong>Chapter {this.props.chapterNum}</strong></h5>
-        <div>{this.props.arrayOfVerses}</div>
-    </div>
-  );
-   }
- });
- var BookTitle = React.createClass({
-   render: function() {
-     return (
-    <h4>{this.props.title} </h4>
-  );
-   }
- });
- var Book = React.createClass({
+ const BookTitle = require('./BookTitle');
+ const Chapter = require('./Chapter');
+ const Verse = require('./Verse');
+
+
+ const Book = React.createClass({
    render: function() {
      var chapterArray = [];
      for (var key in this.props.input) {
@@ -44,12 +21,12 @@
              var verseId = key + ':' + verse;
              var verseText = this.props.input[key][verse];
              arrayOfVerses.push(
-              <Verse id={verseId} verseNumber={verse} verseText={verseText} />
+              <Verse key={verseId} verseNumber={verse} verseText={verseText} />
             );
            }
          }
          chapterArray.push(
-            <Chapter chapterNum={chapterNum} arrayOfVerses={arrayOfVerses}/>
+            <Chapter key={chapterNum} chapterNum={chapterNum} arrayOfVerses={arrayOfVerses}/>
           );
        }
      }
