+++ src/js/fileupload.js
@@ -3,21 +3,27 @@
  *               the more traditional click and open file upload system.
  * @author: Ian Hoegen
  ******************************************************************************/
-const Dropzone = require('react-dropzone');
-const FM = require('./filemodule.js');
-const remote = window.electron.remote;
-const parser = require('./usfm-parse.js');
-const {dialog} = remote;
 const React = require('react');
+
+const Dropzone = require('react-dropzone');
+const FileModule = require('./FileModule');
+
+const remote = window.electron.remote;
+const {dialog} = remote;
+
+const Book = require('./Book');
 const FileActions = require('./FileActions');
-const Book = require('./bibletext');
+
+const parser = require('./usfm-parse.js');
+const style = require('./Style');
+
 var manifestSource = '';
 var bookName = '';
 var joinedChunks = {};
 var currentChapter = '';
 var bookTitle = "";
 
-var FileUploader = React.createClass({
+const FileUploader = React.createClass({
   onDrop: function(files) {
     if (files !== undefined) {
       sendToReader(files[0].path);
@@ -34,25 +40,6 @@ var FileUploader = React.createClass({
   },
 
   render: function() {
-    var style = {
-      dropzone: {
-        width: '100%',
-        color: '#212121',
-        height: '200px',
-        border: '2px dashed #727272',
-        borderRadius: '5px',
-        fontSize: '25px'
-      },
-      dropzoneActive: {
-        border: '2px solid #727272',
-        backgroundColor: '#f5f5f5'
-      },
-      dropzoneText: {
-        lineHeight: '200px',
-        verticalAlign: 'middle',
-        width: '100%'
-      }
-    };
     return (
     <div onClick = {this.onClick} >
       <Dropzone onDrop = {this.onDrop}
@@ -81,7 +68,7 @@ module.exports = FileUploader;
 function sendToReader(file) {
   try {
     manifestSource = file;
-    FM.readFile(file + '/manifest.json', readInManifest);
+    FileModule.readFile(file + '/manifest.json', readInManifest);
     FileActions.setState(false);
   } catch (error) {
     dialog.showErrorBox('Import Error', 'Please make sure that ' +
@@ -100,7 +87,7 @@ function readInManifest(manifest) {
   bookName = bookTitleSplit.join('');
   let bookFileName = bookName + '.json';
   try {
-    FM.readFile('data/ulgb/' + bookFileName, openOriginal);
+    FileModule.readFile('data/ulgb/' + bookFileName, openOriginal);
   } catch (error) {
     console.log(error);
   }
@@ -120,8 +107,8 @@ function readInManifest(manifest) {
 function openUsfmFromChunks(chunk) {
   currentChapter = chunk[0];
   try {
-    FM.readFile(manifestSource + '/' + chunk[0] + '/' + chunk[1] +
-  '.txt', saveChunksLocal);
+    FileModule.readFile(manifestSource + '/' + chunk[0] + '/' + chunk[1] +
+  '.txt', joinChunks);
   } catch (error) {
     dialog.showErrorBox('Import Error', 'Unknown error has occurred');
     console.log(error);
@@ -131,7 +118,7 @@ function openUsfmFromChunks(chunk) {
  * @description This function saves the chunks locally as a window object;
  * @param {string} text - The text being read in from chunks
  ******************************************************************************/
-function saveChunksLocal(text) {
+function joinChunks(text) {
   var currentJoined = joinedChunks;
   if (currentChapter === '00') {
     currentJoined.title = text;
